requirements.txt
attrs==24.2.0
aws-cdk-lib==2.158.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==36.3.0
aws-lambda-powertools==2.43.1
awscli==1.34.22
boto3==1.35.22
botocore==1.35.22
cattrs==23.2.3
colorama==0.4.6
constructs==10.3.0
docutils==0.16
exceptiongroup==1.2.2
importlib_resources==6.4.5
jmespath==1.0.1
jsii==1.103.1
publication==0.0.3
pyasn1==0.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
